bIO/30, Add REST endpoint to index FASTA files
- This is an investigation in the Akka graph DSL. As of now, it works
  but I am not sure why. How is the materialized value chosen?
