+++ stream/src/main/scala/at/bioinform/codec/lucene/LuceneSink.scala
@@ -4,8 +4,6 @@ import akka.stream.stage.{GraphStageLogic, GraphStageWithMaterializedValue, InHa
 import akka.stream.{Attributes, Inlet, SinkShape}
 import at.bioinform.codec.fasta.FastaEntry
 import at.bioinform.lucene.Util
-import org.apache.lucene.analysis.core.WhitespaceAnalyzer
-import org.apache.lucene.analysis.ngram.NGramTokenizer
 import org.apache.lucene.document.Document
 import org.apache.lucene.index.{IndexWriter, IndexWriterConfig}
 import org.apache.lucene.store.Directory
