+++ webapp/conf/routes
@@ -1,5 +1,5 @@
-GET     /                           controllers.HomeController.index
-POST    /                           controllers.HomeController.upload
+GET         /                    controllers.HomeController.index
+POST        /index/add           controllers.LuceneController.add
 
 # Map static resources from the /public folder to the /assets URL path
-GET     /assets/*file               controllers.Assets.versioned(path="/public", file)
+GET         /assets/*file        controllers.Assets.versioned(path="/public", file)
